+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Starting from this version, it will be as backward compatible as possible.
 ## Fixed
 - Removed scaling in the data loader
 - Slide-id passed correctly to dataloader for one-adata multi-slide mode
+- Auto-detect change in n_hops_{local,view} to re-build graph
 
 ## [0.2.4] - 2025-03-26